Add files via upload
Here is a start on the sports activity. I'm trying to create a list view with all razorback athletics where the user cal click on a sport such as "Football" and then move to the next page to pull up the football schedule. I am getting identifier errors that I do not know how to solve. Can anyone find the problem in my code?